requirements.txt
Flask==1.1.1
SQLAlchemy==1.3.10
Werkzeug==0.16.0
Flask_SQLAlchemy==2.4.1
flask_restplus==0.13.0
alembic==1.2.1
Flask_Migrate==2.5.2
python-dotenv==0.10.3
flask_cors==3.0.8
flask_jwt_extended==3.24.1
flask_script==2.0.6
